re-auth update (#19)
* re-auth update

* add default var

* version increase

* add the user email to the config file

in cases where we have to modify the user, we are running into issues if the email is the same. Having it in a centralized spot is going to make it easier to fix some users when they break the user.

* add conditional for user creation

should now report an error when the user fails to create.